Remove xfce stuff; sync with current state.

+++ install-yaourt.sh
@@ -1,11 +1,14 @@
+#!/bin/bash
 
-sudo pacman -Sy --needed base-devel git wget yajl
-git clone https://aur.archlinux.org/package-query.git
-cd package-query/
-makepkg -si
-cd ..
-git clone https://aur.archlinux.org/yaourt.git
-cd yaourt/
-makepkg -si
-cd ..
-rm -rf yaourt package-query
+if [[ ! $(which yaourt) ]]; then
+    sudo pacman -Sy --needed base-devel git wget yajl
+    git clone https://aur.archlinux.org/package-query.git
+    cd package-query/
+    makepkg -si
+    cd ..
+    git clone https://aur.archlinux.org/yaourt.git
+    cd yaourt/
+    makepkg -si
+    cd ..
+    rm -rf yaourt package-query
+fi
